% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai.R
\name{call_oai}
\alias{call_oai}
\title{Call an openapi API}
\usage{
call_oai(path, ..., body = NULL, api_key = Sys.getenv("OPENAI_API_KEY"))
}
\arguments{
\item{path}{The route to an API endpoint. Optionally, a list with the path
plus variables to \code{\link[glue:glue]{glue::glue()}} into the path.}

\item{...}{These dots are for future extensions and must be empty.}

\item{body}{An object to use as the body of the request. If any component of
the body is a path, pass it through \code{\link[fs:path]{fs::path()}} or otherwise give it the
class "fs_path" to indicate that it is a path.}

\item{api_key}{An OpenAI API key.}
}
\value{
The response from the endpoint.
}
\description{
Generate a request to an openapi endpoint.
}
\keyword{internal}
