% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-shared.R
\name{.shared-parameters}
\alias{.shared-parameters}
\title{Parameters used in multiple functions}
\arguments{
\item{title}{The title of the talk.}

\item{description}{(optional) A description of the talk (or any other text you
would like to add to the prompt).}

\item{minutes}{(optional) The length of the talk in minutes.}

\item{section_titles}{A nested list of section titles and (optionally)
minutes per slide as returned by \code{\link[=gen_deck_section_titles]{gen_deck_section_titles()}}, or a
character vector with just the titles.}
}
\description{
Reused parameter definitions are gathered here for easier editing.
}
\keyword{internal}
