% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{extract_section_titles}
\alias{extract_section_titles}
\title{Extract section titles}
\usage{
extract_section_titles(section_titles)
}
\arguments{
\item{section_titles}{A \code{robodeck_section_titles} object.}
}
\value{
A character vector of section titles.
}
\description{
Convert a \code{robodeck_section_titles} object to a character vector of section
titles (removing the "minutes" component).
}
\examples{
# gen_deck_section_titles() returns an object that contains both the section
# titles and their anticipated length.

section_titles_object <- list(
  list(title = "Slide 1", minutes = NULL),
  list(title = "Slide 2", minutes = NULL),
  list(title = "Slide 3", minutes = NULL)
)
extract_section_titles(section_titles_object)
}
