% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{gen_deck}
\alias{gen_deck}
\title{Generate markdown for a revealjs presentation}
\usage{
gen_deck(
  title,
  ...,
  description = NULL,
  minutes = NULL,
  section_titles = NULL,
  outline = NULL,
  additional_information =
    "The tone of the talk should be fun and upbeat. Use an emoji at the start of every bullet in bulleted lists."
)
}
\arguments{
\item{title}{The title of the talk.}

\item{...}{Additional parameters passed on to the OpenAI Chat Completion API.}

\item{description}{(optional) A description of the talk (or any other text you
would like to add to the prompt).}

\item{minutes}{(optional) The length of the talk in minutes.}

\item{section_titles}{A nested list of section titles and (optionally)
minutes per slide as returned by \code{\link[=gen_deck_section_titles]{gen_deck_section_titles()}}, or a
character vector with just the titles.}

\item{outline}{A list of character vectors as returned by
\code{\link[=gen_deck_outline]{gen_deck_outline()}}. The name of each vector is the title of a major
section of the talk, and the vector contains the titles of the slides
within that section.}

\item{additional_information}{(optional) A single string with additional
information about the tone of the talk, specific requirements, etc.}
}
\value{
A single string with markdown to produce slides for a revealjs
presentation.
}
\description{
Generate markdown to produce slides for a conference talk. We strongly
recommend that you use these slides as a starting point, not as the final
slide deck.
}
