% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R
\name{gen_deck_outline}
\alias{gen_deck_outline}
\title{Generate an outline of talk slides}
\usage{
gen_deck_outline(
  title,
  ...,
  api_key = oai_get_default_key(),
  description = NULL,
  minutes = NULL,
  section_titles = NULL
)
}
\arguments{
\item{title}{The title of the talk.}

\item{...}{Additional parameters passed on to the OpenAI Chat Completion API.}

\item{api_key}{An OpenAI API key.}

\item{description}{(optional) A description of the talk (or any other text you
would like to add to the prompt).}

\item{minutes}{(optional) The length of the talk in minutes.}

\item{section_titles}{A nested list of section titles and (optionally)
minutes per slide as returned by \code{\link[=gen_deck_section_titles]{gen_deck_section_titles()}}, or a
character vector with just the titles.}
}
\value{
A list of character vectors. The name of each vector is the title of
a major section of the talk, and the vector contains the titles of the
slides within that section.
}
\description{
Generate a list of slide titles nested inside the major sections of a
conference talk.
}
