% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{gen_deck_section_titles}
\alias{gen_deck_section_titles}
\title{Generate titles for talk sections}
\usage{
gen_deck_section_titles(
  title,
  ...,
  api_key = oai_get_default_key(),
  description = NULL,
  minutes = NULL
)
}
\arguments{
\item{title}{The title of the talk.}

\item{...}{Additional parameters passed on to the OpenAI Chat Completion API.}

\item{api_key}{An OpenAI API key.}

\item{description}{(optional) A description of the talk (or any other text you
would like to add to the prompt).}

\item{minutes}{(optional) The length of the talk in minutes.}
}
\value{
A list of lists, each of which contains a title for a major sections
of the talk, and an empty \code{minutes} object.
}
\description{
Generate a list of titles for the major sections of a conference talk.
}
