% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai.R
\name{oai_get_default_key}
\alias{oai_get_default_key}
\title{Load the default OpenAI API Key}
\usage{
oai_get_default_key(key_name = "OPENAI_API_KEY")
}
\arguments{
\item{key_name}{The name of the keyring service or of the environment
variable.}
}
\value{
A string with the value if available, or \code{""}.
}
\description{
We encourage the use of the keyring package to securely store API keys. This
function attempts to load your OpenAI API key from keyring if it is
available, or from an environment variable otherwise.
}
\examples{
Sys.setenv(TEMP_OAI_API_KEY = "test_value")
test_result <- oai_get_default_key("TEMP_OAI_API_KEY")
test_result
Sys.unsetenv("TEMP_OAI_API_KEY")
}
