% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai.R
\name{oai_set_key}
\alias{oai_set_key}
\title{Set an OpenAI API key}
\usage{
oai_set_key(key_name = "OPENAI_API_KEY", keyring = NULL)
}
\arguments{
\item{key_name}{The name of the key.}

\item{keyring}{The name of a specific keyring, passed on to
\code{keyring::set_key} if available.}
}
\value{
The key, invisibly.
}
\description{
Use the keyring package to store an API key.
}
