% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robodeck-package.R
\docType{package}
\name{robodeck-package}
\alias{robodeck}
\alias{robodeck-package}
\title{robodeck: Quickly Generate Slide Decks with AI}
\description{
Utilize the 'OpenAI' 'API' to generate 'Quarto' slide decks quickly.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/robodeck}
  \item \url{https://jonthegeek.github.io/robodeck/}
  \item Report bugs at \url{https://github.com/jonthegeek/robodeck/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
