% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{robodeck_slide_style}
\alias{robodeck_slide_style}
\title{Description of the overall feel of the slides}
\usage{
robodeck_slide_style()
}
\value{
A string describing the style of the slides.
}
\description{
This is a placeholder for an upcoming, more complex function.
}
\examples{
robodeck_slide_style()
}
