% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{update_section_title_minutes}
\alias{update_section_title_minutes}
\title{Update minutes component of section_titles}
\usage{
update_section_title_minutes(section_titles, section_minutes)
}
\arguments{
\item{section_titles}{A nested list of section titles and (optionally)
minutes per slide as returned by \code{\link[=gen_deck_section_titles]{gen_deck_section_titles()}}, or a
character vector with just the titles.}

\item{section_minutes}{A numeric vector of minutes for each section. This
vector must be the same length as \code{section_titles}.}
}
\value{
An updated list of section titles and their length in minutes.
}
\description{
Update the minutes component of a list of section titles, to assign
approximate times to those sections.
}
