% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai.R
\name{call_oai}
\alias{call_oai}
\title{Call an openapi API}
\usage{
call_oai(path, ..., data = NULL, api_key = Sys.getenv("OPENAI_API_KEY"))
}
\arguments{
\item{path}{The route to an API endpoint.}

\item{...}{These dots are for future extensions and must be empty.}

\item{data}{Data to include in body.}

\item{api_key}{An OpenAI API key.}
}
\value{
The response from the endpoint.
}
\description{
Generate a request to an openapi endpoint.
}
\keyword{internal}
